+++ CHANGELOG.md
@@ -2,6 +2,35 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [3.4.0-alpha.40](https://github.com/uzh-bf/klicker-uzh/compare/v3.4.0-alpha.39...v3.4.0-alpha.40) (2025-10-28)
+
+
+### Features
+
+* add possibility to define point correction for multiple individual students ([#4958](https://github.com/uzh-bf/klicker-uzh/issues/4958)) ([38c92d0](https://github.com/uzh-bf/klicker-uzh/commit/38c92d035956375e28c640b8e90dff207ed1ab96))
+
+## [3.4.0-alpha.39](https://github.com/uzh-bf/klicker-uzh/compare/v3.4.0-alpha.38...v3.4.0-alpha.39) (2025-10-28)
+
+
+### Bug Fixes
+
+* **apps/chat:** add dedicated healthcheck endpoint ([aad5826](https://github.com/uzh-bf/klicker-uzh/commit/aad58263919f5dc3eb8955156411ee2fa42ea9f9))
+
+## [3.4.0-alpha.38](https://github.com/uzh-bf/klicker-uzh/compare/v3.4.0-alpha.37...v3.4.0-alpha.38) (2025-10-15)
+
+
+### Enhancements
+
+* **apps/frontend-manage:** add display name to lecturer cockpit ([#4953](https://github.com/uzh-bf/klicker-uzh/issues/4953)) ([41eefef](https://github.com/uzh-bf/klicker-uzh/commit/41eefefa8d0d53e83254479e1e3cc93b950713c8))
+
+## [3.4.0-alpha.37](https://github.com/uzh-bf/klicker-uzh/compare/v3.4.0-alpha.36...v3.4.0-alpha.37) (2025-10-13)
+
+
+### Bug Fixes
+
+* **cypress:** resolve issues with cypress tests caused by missing element tile in asynchronous activities ([#4952](https://github.com/uzh-bf/klicker-uzh/issues/4952)) ([e5ee90c](https://github.com/uzh-bf/klicker-uzh/commit/e5ee90c430b579018ab4e1d91ed9da4c7e040a98))
+* **packages/shared-components:** handle undefined text for slate conversion ([#4956](https://github.com/uzh-bf/klicker-uzh/issues/4956)) ([80897ce](https://github.com/uzh-bf/klicker-uzh/commit/80897ce1ed510670cd0007939f32ccfcb7d06789))
+
 ## [3.4.0-alpha.36](https://github.com/uzh-bf/klicker-uzh/compare/v3.4.0-alpha.35...v3.4.0-alpha.36) (2025-09-30)